requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
mysql-connector-python==8.2.0 -i https://pypi.tuna.tsinghua.edu.cn/simple
requests==2.31.0 -i https://pypi.tuna.tsinghua.edu.cn/simple
python-dotenv==1.0.0
SQLAlchemy==2.0.25
PyMySQL==1.1.0
Werkzeug==3.0.1
